Add 'NoteProduceService' to separate components interest
Took 3 minutes
