+++ src/app/note/note.module.ts
@@ -19,6 +19,7 @@ import { NotePreviewComponent } from './preview/preview.component';
 import { noteReducerMap } from './reducers';
 import { NoteCollectionSortingMenu } from './shared/note-collection-sorting.menu';
 import { NoteFsService } from './shared/note-fs.service';
+import { NoteProduceService } from './shared/note-produce.service';
 import { NoteViewModeSettingMenu } from './shared/note-view-mode-setting.menu';
 import { NoteWorkspaceComponent } from './workspace/workspace.component';
 
@@ -56,6 +57,7 @@ import { NoteWorkspaceComponent } from './workspace/workspace.component';
         NoteFsService,
         NoteEditorSnippetFactory,
         NoteEditorService,
+        NoteProduceService,
         NoteViewModeSettingMenu,
         NoteCollectionSortingMenu,
     ],
